+++ ft_memmove.c
@@ -6,15 +6,17 @@
 /*   By: inikulin <inikulin@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 19:01:40 by inikulin          #+#    #+#             */
-/*   Updated: 2023/11/23 12:02:43 by inikulin         ###   ########.fr       */
+/*   Updated: 2023/11/24 18:07:23 by inikulin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dest, const void *src, unsigned long long n)
+#include <stddef.h>
+
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	unsigned long long	c;
-	char				*d;
-	const char			*s;
+	size_t		c;
+	char		*d;
+	const char	*s;
 
 	if (n == 0)
 		return (dest);
